Auto stash before merge of "testMaven" and "origin/Items"

+++ SuperPong/Client/src/main/java/view/ItemView.java
@@ -1,21 +1,14 @@
 package view;
 
-import game.Ball;
-import game.Field;
-import javafx.animation.Animation;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
-import javafx.scene.Scene;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
-import javafx.scene.paint.Color;
 import javafx.scene.paint.PhongMaterial;
 import javafx.scene.shape.Box;
 import javafx.scene.transform.Rotate;
 import javafx.util.Duration;
-import main.Displayer;
 import protocole.game.Item;
 
 public class ItemView {
@@ -63,7 +56,7 @@ public class ItemView {
             rotate();
         });
         timeline.getKeyFrames().add(keyFrame);
-        timeline.setCycleCount((int)(10000 / 20)); // stop after 10 seconds
+		timeline.setCycleCount(10000 / 20); // stop after 10 seconds
         timeline.play();
         timeline.setOnFinished(new EventHandler<ActionEvent>() {
             @Override
